test(memory): move mcp.rs tests to a separate file + exclude lock files from Codacy (#1256)
Fixes the two Codacy findings on the 3.4.0 release PR (both 'file too large',
flagged as new only because the release PR's base is main, which lacks the
velesdb-memory crate entirely):

- mcp.rs was 733 lines (616 NLOC) — over the repo's ≤500 standard — almost
  entirely because of a ~440-line inline #[cfg(test)] mod tests. Move it to
  crates/velesdb-memory/src/mcp/server_tests.rs via #[cfg(test)] #[path] mod tests;
  mcp.rs drops to 294 NLOC; the test file matches the repo's existing
  *_tests.rs Codacy exclusion. 27 MCP tests still pass; clippy clean.
- crates/velesdb-node/package-lock.json (1892 NLOC) is a generated lock file, not
  source — add **/package-lock.json and **/Cargo.lock to the global Codacy
  exclude_paths (sdks/typescript/package-lock.json was already excluded).
